Cleanup + ajout pour services sales

+++ src/backend/services/sales/controllers/salesController.js
@@ -1,15 +1,12 @@
-const StoreInventory = require('../models/StoreInventory');
 const Store = require('../models/Store');
 const StoreSale = require('../models/StoreSale');
-const SupplyRequest = require('../models/SupplyRequest');
 const logger = require('../utils/logger');
 
 exports.postNewSaleInStore = async (req, res) => {
   const soldProducts = req.body;
   const storeParam = req.params.storeNumber;
 
-  logger.info("BODY")
-  logger.info(soldProducts);
+  const token = req.headers['authorization'];
 
   const isNumber = !isNaN(parseInt(storeParam)) && parseInt(storeParam) >= 1 && parseInt(storeParam) <= 5;
   const isCentral = storeParam === 'Central';
@@ -22,7 +19,7 @@ exports.postNewSaleInStore = async (req, res) => {
         status: 400,
         error: "Bad Request",
         message: "Numéro de magasin invalide (1-5) ou 'Central'",
-        path: `/api/v1/standard/stores/${storeParam}/sales`
+        path: `/api/v1/sales/stores/${storeParam}`
       }
     );
   }
@@ -39,44 +36,30 @@ exports.postNewSaleInStore = async (req, res) => {
           status: 404,
           error: "Not Found",
           message: `Magasin '${storeName}' introuvable`,
-          path: `/api/v1/standard/stores/${storeParam}/sales`
+          path: `/api/v1/sales/stores/${storeParam}`
         }
       );
     }
 
+    const response = await fetch(`http://krakend:80/api/v1/stocks/stores/${storeParam}/true`, {
+      method: 'PATCH',
+      headers: {
+        'Authorization': token,
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify(soldProducts)
+    });
+
+    if (!response.ok) {
+      const data = await response.json();
+      logger.error(`Erreur retour du stock-service: ${response.status} → ${data.message || JSON.stringify(data)}`);
+      throw new Error(`Stock service responded with ${response.status}: ${data.message}`);
+    }
+
+
     // Gestion des produits
     let totalPrice = 0;
     for (const item of soldProducts) {
-      const product = await StoreInventory.findOne({ store: store._id, name: item.name });
-
-      if (!product) {
-        logger.warn(`Produit '${productName}' introuvable dans le magasin '${storeName}'`);
-        return res.status(404).json(
-          {
-            timestamp: new Date().toISOString(),
-            status: 404,
-            error: "Not Found",
-            message: `Produit '${item.name}' introuvable dans le magasin '${storeParam}'`,
-            path: `/api/v1/standard/stores/${storeParam}/stock/${item.name}`
-          }
-        );
-      }
-
-      if (product.qty < item.qty) {
-        logger.warn(`Stock insuffisant pour '${item.name}'. Disponible : ${product.qty}`);
-        return res.status(400).json(
-          {
-            timestamp: new Date().toISOString(),
-            status: 400,
-            error: "Bad Request",
-            message: `Stock insuffisant pour '${item.name}'. Disponible : ${product.qty}`,
-            path: `/api/v1/standard/stores/${storeParam}/stock/${item.name}`
-          }
-        );
-      }
-
-      product.qty -= item.qty;
-      await product.save();
       totalPrice += item.total_price;
     }
 
@@ -91,14 +74,14 @@ exports.postNewSaleInStore = async (req, res) => {
     res.status(201).json({ message: "Vente enregistrée avec succès." });
 
   } catch (err) {
-    logger.error(`Erreur de communication avec le serveur`);
+    logger.error(`Erreur de communication avec le serveur: ${err}`);
     res.status(500).json(
       {
         timestamp: new Date().toISOString(),
         status: 500,
         error: "Internal Server Error",
         message: "Erreur de communication avec le serveur",
-        path: `/api/v1/standard/stores/${storeParam}/stock/${item.name}`
+        path: `/api/v1/sales/stores/${storeParam}`
       }
     );
   }
@@ -118,7 +101,7 @@ exports.getSalesByStore = async (req, res) => {
         status: 400,
         error: "Bad Request",
         message: "Numéro de magasin invalide (1-5) ou 'Central'",
-        path: `/api/v1/standard/stores/${storeParam}/sales`
+        path: `/api/v1/sales/stores/${storeParam}`
       }
     );
   }
@@ -135,7 +118,7 @@ exports.getSalesByStore = async (req, res) => {
           status: 404,
           error: "Not Found",
           message: `Magasin '${storeName}' introuvable`,
-          path: `/api/v1/standard/stores/${storeParam}/sales`
+          path: `/api/v1/sales/stores/${storeParam}`
         }
       );
     }
@@ -155,14 +138,14 @@ exports.getSalesByStore = async (req, res) => {
     res.status(200).json(sales);
 
   } catch (err) {
-    logger.error(`Erreur de communication avec le serveur`);
+    logger.error(`Erreur de communication avec le serveur: ${err.message}`);
     res.status(500).json(
       {
         timestamp: new Date().toISOString(),
         status: 500,
         error: "Internal Server Error",
         message: "Erreur de communication avec le serveur",
-        path: `/api/v1/standard/stores/${storeParam}/sales`
+        path: `/api/v1/sales/stores/${storeParam}`
       }
     );
   }
@@ -171,6 +154,7 @@ exports.getSalesByStore = async (req, res) => {
 exports.deleteSaleByStore = async (req, res) => {
   const saleId = req.params.saleId;
   const storeParam = req.params.storeNumber;
+  const token = req.headers['authorization'];
 
   const isNumber = !isNaN(parseInt(storeParam)) && parseInt(storeParam) >= 1 && parseInt(storeParam) <= 5;
   const isCentral = storeParam === 'Central';
@@ -183,7 +167,7 @@ exports.deleteSaleByStore = async (req, res) => {
         status: 400,
         error: "Bad Request",
         message: "Numéro de magasin invalide (1-5) ou 'Central'",
-        path: `/api/v1/standard/stores/${storeParam}/sales/${saleId}`
+        path: `/api/v1/sales/stores/${storeParam}/${saleId}`
       }
     );
   }
@@ -200,7 +184,7 @@ exports.deleteSaleByStore = async (req, res) => {
           status: 404,
           error: "Not Found",
           message: `Magasin '${storeName}' introuvable`,
-          path: `/api/v1/standard/stores/${storeParam}/sales/${saleId}`
+          path: `/api/v1/sales/stores/${storeParam}/${saleId}`
         }
       );
     }
@@ -214,21 +198,28 @@ exports.deleteSaleByStore = async (req, res) => {
           status: 404,
           error: "Not Found",
           message: `Vente introuvable dans ce magasin.`,
-          path: `/api/v1/standard/stores/${storeParam}/sales/${saleId}`
+          path: `/api/v1/sales/stores/${storeParam}/${saleId}`
         }
       );
     }
 
-    // Remettre les produits vendus en stock
-    for (const item of sale.contents) {
-      const inventoryItem = await StoreInventory.findOne({ store: store._id, name: item.name });
+    // Appel au service stock pour remettre les produits en stock
+    logger.info(`Retour de vente: envoi PATCH à stocks-service via krakend`);
+    logger.info(`Contenu de la vente à restituer : ${JSON.stringify(sale.contents)}`);
 
-      if (inventoryItem) {
-        inventoryItem.qty += item.qty;
-        await inventoryItem.save();
-      } else {
-        logger.warn(`Produit '${item.name}' non trouvé dans l'inventaire. Impossible de remettre en stock.`);
-      }
+    const response = await fetch(`http://krakend:80/api/v1/stocks/stores/${storeParam}/false`, {
+      method: 'PATCH',
+      headers: {
+        'Authorization': token,
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify(sale.contents)
+    });
+
+    if (!response.ok) {
+      const data = await response.json();
+      logger.error(`Erreur retour du stock-service: ${response.status} → ${data.message || JSON.stringify(data)}`);
+      throw new Error(`Stock service responded with ${response.status}: ${data.message}`);
     }
 
     await StoreSale.deleteOne({ _id: saleId });
@@ -242,7 +233,7 @@ exports.deleteSaleByStore = async (req, res) => {
         status: 500,
         error: "Internal Server Error",
         message: "Erreur de communication avec le serveur",
-        path: `/api/v1/standard/stores/${storeParam}/sales/${saleId}`
+        path: `/api/v1/sales/stores/${storeParam}/${saleId}`
       }
     );
   }
